cmake-git-builder.sh: Use params from env
Use environment variables as in parameters to the script in order to
allow optional parameters. This also improves code by being more
self-documenting in Vagrantfiles (since the environment variable name
gives you a hint of what the parameter is for).

This also adds support for not wiping the source directory of the qmake
project bein built.

WARNING: The default behavior of this script HAS CHANGED where the
source directory is now NOT removed per default. In order to remove the
source directory, please set the WIPE_SRC_DIR parameter to "yes".

Signed-off-by: Jonatan Pålsson <jonatan.palsson@pelagicore.com>