Split bitbake init/fetch scripts + TEMPLATECONF
It is not always the case that we want to do fetchall, so we should
split the two scripts.

This also adds support for setting TEMPLATECONF as an argument to the
init script. However this breaks backward compatibility since one no
longer gives an image as argument to the script.

Signed-off-by: Tobias Olausson <tobias.olausson@pelagicore.com>