shortlog.sh: Take the old manifest as a parameter
Allow comparing against a given manifest file, which might not be part
of the project's git history. This is useful in those cases where the
manifest file contains links to branches and not to git revisions; in
that case, a manifest with proper git revisions can be generated with
the `repo manifest` command.

Also fix a couple of errors, such as the invocation of "repo forall" in
get_changes().

Signed-off-by: Alberto Mardegan <alberto.mardegan@luxoft.com>