runqemu-smoke-test: do not use setcap
setcap has the unfortunate side effect of no longer loading correct
libraries (from qemu sysroot), and instead falling through to host
libraries which may mismatch or be absent altogether.

The reason for adding setcap in the first place is most likely that
someone didn't understand that '1000' is in fact the user/group id
that owns the tun interfaces, and it must match with the user running
the script for qemu to work. So instead of fixing the script to supply
the actual uid/gid that matches tun interfaces, they added a hack to
give qemu the needed capability via setcap.

Signed-off-by: Alexander Kanavin <alex.kanavin@gmail.com>